+++ include/util/sampling/uniform_sampler.h
@@ -28,29 +28,29 @@ class UniformSampler : public ISampler {
   explicit UniformSampler(const QByteArray& data);
   ~UniformSampler() override;
 
-  void setWindowSize(size_t size);
+  void setWindowSize(uint64_t size);
 
  private:
   struct UniformSamplerResampleData : public ResampleData {
-    size_t window_size, windows_count;
-    std::vector<size_t> windows;
+    uint64_t window_size, windows_count;
+    std::vector<uint64_t> windows;
     char* data;
   };
 
   UniformSampler(const UniformSampler& other);
-  char getSampleByte(size_t index) const override;
+  char getSampleByte(uint64_t index) const override;
   const char* getData() const override;
-  size_t getRealSampleSize() const override;
-  size_t getFileOffsetImpl(size_t index) const override;
-  size_t getSampleOffsetImpl(size_t address) const override;
+  uint64_t getRealSampleSize() const override;
+  uint64_t getFileOffsetImpl(uint64_t index) const override;
+  uint64_t getSampleOffsetImpl(uint64_t address) const override;
   ResampleData* prepareResample(SamplerConfig* sc) override;
   void applyResample(ResampleData* rd) override;
   void cleanupResample(ResampleData* rd) override;
   UniformSampler* cloneImpl() const override;
 
-  size_t window_size_, windows_count_;
+  uint64_t window_size_, windows_count_;
   bool use_default_window_size_;
-  std::vector<size_t> windows_;
+  std::vector<uint64_t> windows_;
   char* buffer_;
 };
